PBI #37330 - Get latest stats time. Fill with null.

+++ Domain/Intraday/IncomingIntervalModel.cs
@@ -2,7 +2,7 @@ namespace Teleopti.Ccc.Domain.Intraday
 {
 	public class IncomingIntervalModel
 	{
-		public double IntervalId { get; set; }
+		public int IntervalId { get; set; }
 		public double ForecastedCalls { get; set; }
 		public double ForecastedHandleTime { get; set; }
 		public double ForecastedAverageHandleTime { get; set; }
